Move all decorators to `decorators.py`
